Split HIR symbol helpers
Move monomorphized symbol naming and async/intrinsic classification out of the main HIR lowering facade into a tracked compiler.hir child module.

Lower the compiler source monolith ratchet for hir.rs, top-file lines, and top-file share; add symbols.rs to the report fixture.

Refs #1254.
